Move validate to a file
