+++ src/config_paths.rs
@@ -9,6 +9,17 @@ lazy_static! {
 
 pub static CONFIG_PATHS: OnceCell<Vec<PathBuf>> = OnceCell::new();
 
+/// Expands, dedups, and sets global values.
+pub fn prepare(paths: Vec<PathBuf>) -> Option<Vec<PathBuf>> {
+    let mut config_paths = expand(paths)?;
+    config_paths.sort();
+    config_paths.dedup();
+    CONFIG_PATHS
+        .set(config_paths.clone())
+        .expect("Cannot set global config paths");
+    Some(config_paths)
+}
+
 /// Expand a list of paths (potentially containing glob patterns) into real
 /// config paths, replacing it with the default paths when empty.
 pub fn expand(config_paths: Vec<PathBuf>) -> Option<Vec<PathBuf>> {
