+++ src/lib.rs
@@ -37,6 +37,7 @@ pub mod trace;
 pub mod transforms;
 pub mod types;
 pub mod unit_test;
+pub mod validate;
 
 pub use event::Event;
